refactor hints phase 1, making the HintSet
