+++ archipelago/Hints.py
@@ -1,11 +1,10 @@
 """Hints for DK64R Archipelago."""
 
 # from worlds.dk64 import DK64World
-from randomizer.CompileHints import UpdateSpoilerHintList, getRandomHintLocation, replaceKongNameWithKrusha
+from randomizer.CompileHints import HintSet, UpdateSpoilerHintList, replaceKongNameWithKrusha
 from randomizer.Enums.Maps import Maps
 from randomizer.Enums.Kongs import Kongs
 from randomizer.Enums.Levels import Levels
-from randomizer.Lists.WrinklyHints import ClearHintMessages
 from randomizer.Patching.UpdateHints import UpdateHint
 
 boss_names = {
@@ -106,7 +105,7 @@ def convert_hint_door_name_to_full_name(hint_door_name):
 def CompileArchipelagoHints(world, hint_data: list):
     """Insert Archipelago hints."""
     replaceKongNameWithKrusha(world.spoiler)
-    ClearHintMessages()
+    hintset = HintSet()
     hint_location_mapping = {}
     if world.options.hint_style == 1:
         woth_count = 0  # disabled
@@ -119,7 +118,7 @@ def CompileArchipelagoHints(world, hint_data: list):
 
     # Variables
     hints_remaining = 35  # Keep count how many hints we placed
-    hints = []  # The hints we compile
+    hintset.hints = []  # The hints we compile
     hint_locations_used = []  # Track which hint locations are used for mapping
     woth_duplicates = []
     kong_locations = hint_data["kong"]
@@ -130,21 +129,21 @@ def CompileArchipelagoHints(world, hint_data: list):
     already_hinted = kong_locations + key_locations
     hint_location_pairs = []
     krool_hint = parseKRoolHint(world)
-    hints.append(krool_hint)
+    hintset.hints.append(krool_hint)
     hint_location_pairs.append((krool_hint, None))  # K. Rool hints don't have a specific location
     hints_remaining -= 1
 
     # Kong hints
     for kong_loc in kong_locations:
         kong_hint = parseKongHint(world, kong_loc)
-        hints.append(kong_hint)
+        hintset.hints.append(kong_hint)
         hint_location_pairs.append((kong_hint, kong_loc))
         hints_remaining -= 1
 
     # Key hints
     for key_loc in key_locations:
         key_hint = parseKeyHint(world, key_loc)
-        hints.append(key_hint)
+        hintset.hints.append(key_hint)
         hint_location_pairs.append((key_hint, key_loc))
         hints_remaining -= 1
 
@@ -155,7 +154,7 @@ def CompileArchipelagoHints(world, hint_data: list):
     for woth_loc in woth_locations:
         already_hinted.append(woth_loc)
         this_hint = parseWothHint(world, woth_loc)
-        hints.append(this_hint)
+        hintset.hints.append(this_hint)
         hint_location_pairs.append((this_hint, None))
         woth_duplicates.append(this_hint)
         hints_remaining -= 1
@@ -166,7 +165,7 @@ def CompileArchipelagoHints(world, hint_data: list):
     major_locations = world.spoiler.settings.random.sample(major_locations, major_count)
     for major_loc in major_locations:
         major_hint = parseMajorItemHint(world, major_loc)
-        hints.append(major_hint)
+        hintset.hints.append(major_hint)
         hint_location_pairs.append((major_hint, major_loc))
         hints_remaining -= 1
 
@@ -175,14 +174,14 @@ def CompileArchipelagoHints(world, hint_data: list):
     deep_locations = world.spoiler.settings.random.sample(deep_locations, deep_count)
     for deep_loc in deep_locations:
         deep_hint = parseDeepHint(world, deep_loc)
-        hints.append(deep_hint)
+        hintset.hints.append(deep_hint)
         hint_location_pairs.append((deep_hint, deep_loc))
         hints_remaining -= 1
 
     # Woth hint duplicates as needed
     while hints_remaining > 0 and len(woth_duplicates) > 0:
         duplicate_hint = woth_duplicates.pop()
-        hints.append(duplicate_hint)
+        hintset.hints.append(duplicate_hint)
         hint_location_pairs.append((duplicate_hint, None))  # Duplicates don't need location mapping
         hints_remaining -= 1
 
@@ -192,13 +191,13 @@ def CompileArchipelagoHints(world, hint_data: list):
         print("Not enough hints. Please wait. stage_generate_output might be crashing.")
         while hints_remaining > 0:
             filler_hint = "no hint, sorry...".upper()
-            hints.append(filler_hint)
+            hintset.hints.append(filler_hint)
             hint_location_pairs.append((filler_hint, None))  # Filler hints don't have locations
             hints_remaining -= 1
 
     # Place hints and create mapping
     for i, (hint, location_obj) in enumerate(hint_location_pairs):
-        hint_location = getRandomHintLocation(random=world.spoiler.settings.random)
+        hint_location = hintset.getRandomHintLocation(random=world.spoiler.settings.random)
 
         if hint_location is None:
             continue
@@ -233,7 +232,7 @@ def CompileArchipelagoHints(world, hint_data: list):
                     # Fallback if import fails - hints will not work but generation can continue
                     pass
 
-    UpdateSpoilerHintList(world.spoiler)
+    UpdateSpoilerHintList(world.spoiler, hintset)
 
     # Store the hint location mapping in the world for use in slot_data
     world.hint_location_mapping = hint_location_mapping
